requirements.txt
# Core dependencies
streamlit==1.28.1
pandas==2.1.3
numpy==1.26.2

# Tableau file parsing
lxml==4.9.3
xmltodict==0.13.0

# Snowflake connection
snowflake-connector-python==3.5.0
snowflake-sqlalchemy==1.5.0

# Visualization libraries
plotly==5.18.0
altair==5.1.2
matplotlib==3.8.2
seaborn==0.13.0

# AI and computer vision
anthropic==0.7.1
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.22.0

# Utilities
python-dotenv==1.0.0
jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.6

# File handling
openpyxl==3.1.2
xlrd==2.0.1

# Deployment
gunicorn==21.2.0